+++ src/node/test/historical_queries.cpp
@@ -397,7 +397,7 @@ TEST_CASE("StateCache point queries")
 
     {
       INFO("Dropping a handle deletes it, and it can no longer be retrieved");
-      cache.drop_request(default_handle);
+      cache.drop_cached_states(default_handle);
       const auto state =
         cache.get_state_at(default_handle, high_signature_transaction);
       REQUIRE(state == nullptr);
@@ -505,19 +505,19 @@ TEST_CASE("StateCache get store vs get state")
     REQUIRE(cache.get_store_at(default_handle, seqno_a) == nullptr);
     REQUIRE(provide_ledger_entry(seqno_a));
     REQUIRE(cache.get_store_at(default_handle, seqno_a) != nullptr);
-    cache.drop_request(default_handle);
+    cache.drop_cached_states(default_handle);
 
     REQUIRE(cache.get_store_at(default_handle, seqno_b) == nullptr);
     REQUIRE(provide_ledger_entry(seqno_b));
     REQUIRE(cache.get_store_at(default_handle, seqno_b) != nullptr);
-    cache.drop_request(default_handle);
+    cache.drop_cached_states(default_handle);
 
     REQUIRE(
       cache.get_store_at(default_handle, signature_transaction) == nullptr);
     REQUIRE(provide_ledger_entry(signature_transaction));
     REQUIRE(
       cache.get_store_at(default_handle, signature_transaction) != nullptr);
-    cache.drop_request(default_handle);
+    cache.drop_cached_states(default_handle);
   }
 
   {
@@ -529,7 +529,7 @@ TEST_CASE("StateCache get store vs get state")
     auto state_a = cache.get_state_at(default_handle, seqno_a);
     REQUIRE(state_a != nullptr);
     REQUIRE(state_a->receipt != nullptr);
-    cache.drop_request(default_handle);
+    cache.drop_cached_states(default_handle);
 
     REQUIRE(cache.get_state_at(default_handle, seqno_b) == nullptr);
     REQUIRE(provide_ledger_entry(seqno_b));
@@ -538,7 +538,7 @@ TEST_CASE("StateCache get store vs get state")
     auto state_b = cache.get_state_at(default_handle, seqno_b);
     REQUIRE(state_b != nullptr);
     REQUIRE(state_b->receipt != nullptr);
-    cache.drop_request(default_handle);
+    cache.drop_cached_states(default_handle);
 
     REQUIRE(
       cache.get_state_at(default_handle, signature_transaction) == nullptr);
@@ -546,7 +546,7 @@ TEST_CASE("StateCache get store vs get state")
     auto state_sig = cache.get_state_at(default_handle, signature_transaction);
     REQUIRE(state_sig != nullptr);
     REQUIRE(state_sig->receipt != nullptr);
-    cache.drop_request(default_handle);
+    cache.drop_cached_states(default_handle);
   }
 
   {
@@ -561,7 +561,7 @@ TEST_CASE("StateCache get store vs get state")
       auto state_a = cache.get_state_at(default_handle, seqno_a);
       REQUIRE(state_a != nullptr);
       REQUIRE(state_a->receipt != nullptr);
-      cache.drop_request(default_handle);
+      cache.drop_cached_states(default_handle);
     }
 
     {
@@ -576,7 +576,7 @@ TEST_CASE("StateCache get store vs get state")
       state_b = cache.get_state_at(default_handle, seqno_b);
       REQUIRE(state_b != nullptr);
       REQUIRE(state_b->receipt != nullptr);
-      cache.drop_request(default_handle);
+      cache.drop_cached_states(default_handle);
     }
 
     {
@@ -590,7 +590,7 @@ TEST_CASE("StateCache get store vs get state")
         cache.get_state_at(default_handle, signature_transaction);
       REQUIRE(state_sig != nullptr);
       REQUIRE(state_sig->receipt != nullptr);
-      cache.drop_request(default_handle);
+      cache.drop_cached_states(default_handle);
     }
   }
 }