+++ tests/infra/node.py
@@ -446,10 +446,7 @@ class Node:
             )
 
     def session_auth(self, name=None):
-        return {
-            "session_auth": self.identity(name),
-            "ca": os.path.join(self.common_dir, "networkcert.pem"),
-        }
+        return {"session_auth": self.identity(name)}
 
     def signing_auth(self, name=None):
         return {"signing_auth": self.identity(name)}
@@ -457,14 +454,27 @@ class Node:
     def get_public_rpc_host(self):
         return self.remote.get_host()
 
+    def session_ca(self, self_signed_ok):
+        if self_signed_ok:
+            return {"ca": ""}
+        else:
+            return {"ca": os.path.join(self.common_dir, "networkcert.pem")}
+
     def client(
-        self, identity=None, signing_identity=None, interface_idx=None, **kwargs
+        self,
+        identity=None,
+        signing_identity=None,
+        interface_idx=None,
+        self_signed_ok=False,
+        **kwargs,
     ):
         if self.network_state == NodeNetworkState.stopped:
             raise RuntimeError(
                 f"Cannot create client for node {self.local_node_id} as node is stopped"
             )
-        akwargs = self.session_auth(identity)
+
+        akwargs = self.session_ca(self_signed_ok)
+        akwargs.update(self.session_auth(identity))
         akwargs.update(self.signing_auth(signing_identity))
         akwargs[
             "description"