+++ tests/nobuiltins.py
@@ -10,33 +10,16 @@ import time
 def test_nobuiltins_endpoints(network, args):
     primary, backups = network.find_nodes()
     with primary.client() as c:
-        timeout = 3
-        end_time = time.time() + timeout
-        found_stable_commit = False
-        r = c.get("/node/network")
+        r = c.get("/app/commit")
         assert r.status_code == HTTPStatus.OK
-        target_view = r.view
-        target_seqno = r.seqno
-        assert target_view is not None
-        assert target_seqno is not None
-        while time.time() < end_time:
-            r = c.get("/node/commit")
-            assert r.status_code == HTTPStatus.OK
-            body_j = r.body.json()
-            tx_id = TxID.from_str(body_j["transaction_id"])
-            if tx_id.view == target_view and tx_id.seqno == target_seqno:
-                found_stable_commit = True
-                break
-            else:
-                time.sleep(0.1)
-
-        assert found_stable_commit, f"Failed to reach stable commit after {timeout}s"
+        body_j = r.body.json()
+        tx_id = TxID.from_str(body_j["transaction_id"])
 
         r = c.get("/app/node_summary")
         assert r.status_code == HTTPStatus.OK
         body_j = r.body.json()
-        assert body_j["committed_view"] == tx_id.view
-        assert body_j["committed_seqno"] == tx_id.seqno
+        assert body_j["committed_view"] >= tx_id.view
+        assert body_j["committed_seqno"] >= tx_id.seqno
         assert body_j["quote_format"] == "OE_SGX_v1"
         assert body_j["node_id"] == primary.node_id