+++ tests/reconfiguration.py
@@ -8,12 +8,13 @@ import suite.test_requirements as reqs
 import tempfile
 from shutil import copy
 import os
-from infra.checker import check_can_progress, check_does_not_progress
+import time
 import ccf.ledger
 import json
 import infra.crypto
 from datetime import datetime
-
+from infra.checker import check_can_progress
+from infra.runner import ConcurrentRunner
 
 from loguru import logger as LOG
 
@@ -36,6 +37,33 @@ def count_nodes(configs, network):
         nodes_in_config = set(node_config.keys()) - stopped
         assert nodes == nodes_in_config, f"{nodes} {nodes_in_config} {node_id}"
     return len(nodes)
+
+
+def wait_for_reconfiguration_to_complete(network, timeout=10):
+    max_num_configs = 0
+    max_rid = 0
+    all_same_rid = False
+    end_time = time.time() + timeout
+    while max_num_configs > 1 or not all_same_rid:
+        max_num_configs = 0
+        all_same_rid = True
+        for node in network.nodes:
+            with node.client(self_signed_ok=True) as c:
+                try:
+                    r = c.get("/node/consensus")
+                    rj = r.body.json()
+                    cfgs = rj["details"]["configs"]
+                    num_configs = len(cfgs)
+                    max_num_configs = max(max_num_configs, num_configs)
+                    if num_configs == 1 and cfgs[0]["rid"] != max_rid:
+                        max_rid = max(max_rid, cfgs[0]["rid"])
+                        all_same_rid = False
+                except Exception as ex:
+                    # OK, retiring node may be gone or a joining node may not be ready yet
+                    LOG.info(f"expected RPC failure because of: {ex}")
+        LOG.info(f"max num configs: {max_num_configs}, max rid: {max_rid}")
+        if time.time() > end_time:
+            raise Exception("Reconfiguration did not complete in time")
 
 
 @reqs.description("Adding a valid node without snapshot")
@@ -189,6 +217,14 @@ def test_add_as_many_pending_nodes(network, args):
 
     for new_node in new_nodes:
         network.retire_node(primary, new_node)
+
+    wait_for_reconfiguration_to_complete(network)
+
+    # Stop the retired nodes so they don't linger in the background and interfere
+    # with subsequent tests
+    for new_node in new_nodes:
+        new_node.stop()
+
     return network
 
 
@@ -201,6 +237,7 @@ def test_retire_backup(network, args):
     network.retire_node(primary, backup_to_retire)
     backup_to_retire.stop()
     check_can_progress(primary)
+    wait_for_reconfiguration_to_complete(network)
     return network
 
 
@@ -210,7 +247,7 @@ def test_retire_primary(network, args):
     pre_count = count_nodes(node_configs(network), network)
 
     primary, backup = network.find_primary_and_any_backup()
-    network.retire_node(primary, primary)
+    network.retire_node(primary, primary, timeout=15)
     # Query this backup to find the new primary. If we ask any other
     # node, then this backup may not know the new primary by the
     # time we call check_can_progress.
@@ -219,6 +256,7 @@ def test_retire_primary(network, args):
     post_count = count_nodes(node_configs(network), network)
     assert pre_count == post_count + 1
     primary.stop()
+    wait_for_reconfiguration_to_complete(network)
     return network
 
 
@@ -270,6 +308,7 @@ def test_version(network, args):
 
 
 @reqs.description("Replace a node on the same addresses")
+@reqs.can_kill_n_nodes(1)
 def test_node_replacement(network, args):
     primary, backups = network.find_nodes()
 
@@ -351,6 +390,7 @@ def test_join_straddling_primary_replacement(network, args):
 
     primary.stop()
     network.nodes.remove(primary)
+    wait_for_reconfiguration_to_complete(network)
     return network
 
 
@@ -388,14 +428,24 @@ def test_retiring_nodes_emit_at_most_one_signature(network, args):
     assert not retiring_nodes, (retiring_nodes, retired_nodes)
     LOG.info("{} nodes retired throughout test", len(retired_nodes))
 
+    wait_for_reconfiguration_to_complete(network)
+
     return network
 
 
 @reqs.description("Adding a learner without snapshot")
 def test_learner_catches_up(network, args):
-    args.join_timer = args.join_timer * 2
+    primary, _ = network.find_primary()
+    num_nodes_before = 0
 
-    num_nodes_before = len(network.nodes)
+    with primary.client() as c:
+        s = c.get("/node/consensus")
+        rj = s.body.json()
+        # At this point, there should be exactly one configuration
+        assert len(rj["details"]["configs"]) == 1
+        c0 = rj["details"]["configs"][0]["nodes"]
+        num_nodes_before = len(c0)
+
     new_node = network.create_node("local://localhost")
     network.join_node(new_node, args.package, args, from_snapshot=False)
     network.trust_node(new_node, args)
@@ -405,7 +455,6 @@ def test_learner_catches_up(network, args):
         rj = s.body.json()
         assert rj["status"] == "Learner" or rj["status"] == "Trusted"
 
-    primary, _ = network.find_primary()
     network.consortium.wait_for_node_to_exist_in_store(
         primary,
         new_node.node_id,
@@ -419,34 +468,11 @@ def test_learner_catches_up(network, args):
         assert len(rj["details"]["learners"]) == 0
 
         # At this point, there should be exactly one configuration, which includes the new node.
-        print(rj)
         assert len(rj["details"]["configs"]) == 1
         c0 = rj["details"]["configs"][0]["nodes"]
-        assert len(c0) == num_nodes_before + 1 and new_node.node_id in c0
+        assert len(c0) == num_nodes_before + 1
+        assert new_node.node_id in c0
 
-    return network
-
-
-@reqs.description("Add a learner, suspend nodes, check that there is no progress")
-def test_learner_does_not_take_part(network, args):
-    primary, backups = network.find_nodes()
-    nodes = network.get_joined_nodes()
-    f = infra.e2e_args.max_f(args, len(nodes)) + 1
-    f_backups = backups[:f]
-
-    new_node = network.create_node("local://localhost")
-    network.join_node(new_node, args.package, args, from_snapshot=False)
-    network.trust_node(new_node, args)
-
-    # No way to keep the new node suspended in learner state?
-
-    for b in f_backups:
-        b.suspend()
-    check_does_not_progress(primary)
-    primary_after, _ = network.find_primary()
-    for b in f_backups:
-        b.resume()
-    assert primary.node_id == primary_after.node_id
     return network
 
 
@@ -502,10 +528,10 @@ def run(args):
 
             test_node_filter(network, args)
             test_retiring_nodes_emit_at_most_one_signature(network, args)
-        else:
+
+        if args.reconfiguration_type == "2tx":
             test_learner_catches_up(network, args)
-            # test_learner_does_not_take_part(network, args)
-            test_retire_backup(network, args)
+
         test_node_certificates_validity_period(network, args)
         test_add_node_invalid_validity_period(network, args)
 
@@ -570,14 +596,38 @@ def run_join_old_snapshot(args):
                 pass
 
 
+def run_all(args):
+    run(args)
+    if cr.args.consensus != "bft":
+        run_join_old_snapshot(all)
+
+
 if __name__ == "__main__":
 
-    args = infra.e2e_args.cli_args()
-    args.package = "samples/apps/logging/liblogging"
-    args.nodes = infra.e2e_args.min_nodes(args, f=1)
-    args.initial_user_count = 1
+    def add(parser):
+        parser.add_argument(
+            "--include-2tx-reconfig",
+            help="Include tests for the 2-transaction reconfiguration scheme",
+            action="store_true",
+        )
 
-    run(args)
+    cr = ConcurrentRunner(add)
 
-    if args.consensus != "bft":
-        run_join_old_snapshot(args)
+    cr.add(
+        "1tx_reconfig",
+        run,
+        package="samples/apps/logging/liblogging",
+        nodes=infra.e2e_args.min_nodes(cr.args, f=1),
+        reconfiguration_type="1tx",
+    )
+
+    if cr.args.include_2tx_reconfig:
+        cr.add(
+            "2tx_reconfig",
+            run,
+            package="samples/apps/logging/liblogging",
+            nodes=infra.e2e_args.min_nodes(cr.args, f=1),
+            reconfiguration_type="2tx",
+        )
+
+    cr.run()